Feat: Implements authentication guard
Adds an authentication guard to protect specific routes,
redirecting unauthenticated users to the welcome page.
Additionally, it ensures `localStorage` interactions only
occur in browser environments to prevent server-side errors.

+++ src/app/footer/footer.component.ts
@@ -1,10 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, Inject, PLATFORM_ID } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { Router, RouterModule } from '@angular/router';
+import { isPlatformBrowser } from '@angular/common';
 
 @Component({
   selector: 'app-footer',
@@ -23,13 +24,21 @@ import { Router, RouterModule } from '@angular/router';
 export class FooterComponent {
   today = new Date();
 
-  constructor(private router: Router) {}
+  constructor(
+    private router: Router,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {}
 
   get isLoggedIn(): boolean {
-    return !!localStorage.getItem('user') && !!localStorage.getItem('token');
+    if (isPlatformBrowser(this.platformId)) {
+      return !!localStorage.getItem('user') && !!localStorage.getItem('token');
+    }
+    return false;
   }
   userLogout(): void {
-    localStorage.clear();
+    if (isPlatformBrowser(this.platformId)) {
+      localStorage.clear();
+    }
     this.router.navigate(['/welcome']);
   }
 }
